link from account name to first role console
This is great for things like vimperator since it's possible to do
something like 'click the link my-account' to log in without the mouse.
Without this I have to do something like 'click on the 27th AWSConsole
link'...

The link will point to the first role name (I don't know if this should
be made more clever eG always link to the full-access role or
equivalent).
But just using the first role feels right. In our production AFP we have
exactly 2 accounts with 2 roles, everything else just has one role.